fix: log cache fallbacks
This change is intended to help troubleshoot a caching issue raised here:
https://github.com/veepee-oss/terustry/issues/1

WARN messages are added to indicate a fetch failed, either for artifacts
or for versions, and the previous Ok result was used instead.

A timeout of 10 seconds is also added to the reqwest client.

Recent changes were added in as well.